fix: async QJO execution
This commit refactors the method used to optimize JavaScript files from
'optimizeJavaScriptViaQuiqqerJO' to 'optimizeJavaScriptViaQJO'. The references made to this method
in 'Cache/Handler.php' and 'Cache/Optimizer.php' files have also been updated. In the
'Optimizer.php', additional documentation has been added for clarity. This change will ensure that
the 'QJO' method is consistently referred to across the code base.

Additions were also made to 'locale.xml', a notice stating that the function is in the testing
phase.